recommendation/requirements.txt
Flask
numpy
pandas
scikit-learn
SQLAlchemy
mysql-connector-python
python-dotenv